extract emptyState
